==> 1_preprocessing.ipynb <==
# ОПИСАНИЕ ЗАДАЧИ

Данные:
- чековые данные (transactions.parquet, для чтения через pandas дополнительно нужно установить библиотеку pyarrow)
- справочник товаров (materials.csv)
- справочник магазинов (plants.csv)
- справочник клиентов (clients.csv)
Более подробное описание данных дано в файле Data Description.

Цель: 
1) проанализировать данные и определить оптимальную методологию определения отточных клиентов
2) разработать модель вероятности оттока клиентов по выбранной вами методологии
3) дать интерпретацию разработанной модели, ответить на вопросы: какие признаки наиболее влияют на отток клиентов


# ЧАСТЬ 1. ПРЕДОБРАБОТКА ВХОДНЫХ ДАННЫХ

### Загрузка библиотек

import pyarrow.parquet as pq
import pandas as pd
from pathlib import Path
import numpy as np
import pickle

# 1. Загрузка данных

workdir = str(Path().absolute())

transactions = pq.read_table(workdir+'/transactions.parquet')
transactions = transactions.to_pandas()
clients = pd.read_csv(workdir+'/clients.csv')
materials = pd.read_csv(workdir+'/materials.csv')
plants = pd.read_csv(workdir+'/plants.csv')

Посмотрим на данные.

transactions.info()

clients.info()

materials.info()

plants.info()

print(80*'*')
print('transactions:')
display(transactions.head())

print(80*'*')
print('clients:')
display(clients.head())

print(80*'*')
print('materials:')
display(materials.head())

print(80*'*')
print('plants:')
display(plants.head())

Выводы:
* данные transactions занимают очень много памяти
* для некоторых данных возможно уменьшение размера потребляемой памяти
* данные clients содержат пропущенные значения в столбцах gender и birthyear

# 2. Предобработка данных

### Уменьшение размера уникальных ключей

Посмотрим на количество уникальных ключей в каждом датафрейме.

print('transactions:')
print(f'     Уникальных chq_id: {transactions.chq_id.nunique()}')
print(f'     Уникальных client_id: {transactions.client_id.nunique()}')
print(f'     Уникальных material: {transactions.material.nunique()}')
print(f'     Уникальных plant: {transactions.plant.nunique()}')
print('clients:')
print(f'     Уникальных client_id: {clients.client_id.nunique()}')
print('materials:')
print(f'     Уникальных material: {materials.material.nunique()}')
print('plants:')
print(f'     Уникальных plant: {plants.plant.nunique()}')

Уникальных ключей намного меньше, чем позволяет хранить их длинный шестнадцатиричный формат. В рамках данного анализа целесообразно перевести эти ключи в числовой формат меньшей размерности. Это позволит существенно уменьшить потребление памяти.

client_ids = pd.DataFrame({'client_id':transactions.client_id.drop_duplicates()\
                                                       .append(clients.client_id.drop_duplicates())\
                                                       .drop_duplicates()})
client_ids['num'] = client_ids['client_id'].astype('category').cat.codes

material_ids = pd.DataFrame({'material':transactions.material.drop_duplicates()\
                                                       .append(materials.material.drop_duplicates())\
                                                       .drop_duplicates()})
material_ids['num'] = material_ids['material'].astype('category').cat.codes

plant_ids = pd.DataFrame({'plant':transactions.plant.drop_duplicates()\
                                                       .append(plants.plant.drop_duplicates())\
                                                       .drop_duplicates()})
plant_ids['num'] = plant_ids['plant'].astype('category').cat.codes

chq_ids = pd.DataFrame({'chq_id':transactions.chq_id.drop_duplicates()})
chq_ids['num'] = chq_ids['chq_id'].astype('category').cat.codes

transactions = transactions.merge(client_ids, how='left', on='client_id')\
                           .drop(columns='client_id')\
                           .rename(columns={'num':'client_id'})
transactions = transactions.merge(material_ids, how='left', on='material')\
                           .drop(columns='material')\
                           .rename(columns={'num':'material'})
transactions = transactions.merge(plant_ids, how='left', on='plant')\
                           .drop(columns='plant')\
                           .rename(columns={'num':'plant'})
transactions = transactions.merge(chq_ids, how='left', on='chq_id')\
                           .drop(columns='chq_id')\
                           .rename(columns={'num':'chq_id'})
clients = clients.merge(client_ids, how='left', on='client_id')\
                           .drop(columns='client_id')\
                           .rename(columns={'num':'client_id'})
materials = materials.merge(material_ids, how='left', on='material')\
                           .drop(columns='material')\
                           .rename(columns={'num':'material'})
plants = plants.merge(plant_ids, how='left', on='plant')\
                           .drop(columns='plant')\
                           .rename(columns={'num':'plant'})

### Предобработка transactions

Посмотрим на object-данные по позиции в чеке (chq_position).

transactions['chq_position'].unique()

Все данные являются числовыми. Убедимся, что данные формата "00000000XX" эквиваленты данным формата "0XX".

transactions[transactions.chq_position=='0000000010'].head()

transactions[transactions['chq_id'] == 106596].sort_values(by='chq_position')

Переведем chq_position в формат int.

transactions.chq_position = transactions.chq_position.astype(int)

Посмотрим на данные is_promo.

transactions.is_promo.unique()

Переведем их в boolean-формат.

transactions.is_promo = transactions.is_promo.astype(bool)

Попробуем уменьшить размер типов данных float.

for col in ['sales_count', 'sales_sum']:
    c_min = transactions[col].min()
    c_max = transactions[col].max()
    if c_min > np.finfo(np.float16).min and c_max < np.finfo(np.float16).max:
        transactions[col] = transactions[col].astype(np.float16)
    elif c_min > np.finfo(np.float32).min and c_max < np.finfo(np.float32).max:
        transactions[col] = transactions[col].astype(np.float32)
    else:
        transactions[col] = transactions[col].astype(np.float64)  

transactions.info()

### Предобработка clients

Посмотрим на данные birthyear.

clients.birthyear.unique()

clients.birthyear.min()

Временно заменим nan-значения birthyear на 1899 и преобразуем столбец в int-формат.

clients.birthyear.fillna(1899., inplace=True)
clients.birthyear = clients.birthyear.astype(int)

Посмотрим на данные gender.

clients.gender.unique()

Временно заменим nan-значения gender на 0 и преобразуем столбец в int-формат.

clients.gender.fillna(0, inplace=True)
clients.gender = clients.gender.astype('category').cat.codes

Посмотрим на данные city.

clients.city.unique()

Преобразуем столбец в категориальный формат. Категории городов также потребуются для преобразования данных plants.

city_ids = pd.DataFrame({'city':clients.city.drop_duplicates()})
city_ids['num'] = city_ids['city'].astype('category').cat.codes
clients = clients.merge(city_ids, how='left', on='city')\
                           .drop(columns='city')\
                           .rename(columns={'num':'city'})

Проверим в transactions наличие данных по всем пользователям:

missing_clients = transactions[~transactions.client_id.isin(clients.client_id)].client_id.unique()
print(f"Клиенты с неизвестными данными:{missing_clients}")
print(f"Количество транзакций неизвестных клиентов:{transactions[transactions.client_id.isin(missing_clients)].client_id.count()}")

Учитывая очень маленькое количество неизвестных клиентов и записей по ним, удалим их из данных.

transactions = transactions[transactions.client_id.isin(clients.client_id)]

Для задачи потребуется информация о стоимости единицы товара. Для этого добавим дополнительный столбец price (стоимость единицы товара).

transactions['price'] = transactions['sales_sum'] / transactions['sales_count']

Результат обработки:

transactions.info()

clients.info()

### Предобработка materials

Преобразуем столбцы с подкатегориями товара и производителями в категориальный формат.

materials.hier_level_1 = materials.hier_level_1.astype('category').cat.codes
materials.hier_level_2 = materials.hier_level_2.astype('category').cat.codes
materials.hier_level_3 = materials.hier_level_3.astype('category').cat.codes
materials.hier_level_4 = materials.hier_level_4.astype('category').cat.codes

materials.vendor = materials.vendor.astype('category').cat.codes

Посмотрим на столбцы is_private_label, is_alco.

materials.is_private_label.unique()

materials.is_alco.unique()

Преобразуем их в формат boolean.

materials.is_private_label = materials.is_private_label.astype(bool)
materials.is_alco = materials.is_alco.astype(bool)

Проверим в transactions наличие данных по всем товарам:

missing_materials = transactions[~transactions.material.isin(materials.material)].material.unique()
print(f"Товары с неизвестными данными:{missing_materials}")
print(f"Количество транзакций с неизвестными товарами:{transactions[transactions.material.isin(missing_materials)].material.count()}")

Учитывая очень маленькое количество неизвестных товаров и записей по ним, удалим их из данных.

transactions = transactions[transactions.material.isin(materials.material)]

Результат обработки:

transactions.info()

materials.info()

### Предобработка plants

Посмотрим на данные plant_type.

plants.plant_type.unique()

Преобразуем их в категориальный формат

plants.plant_type = plants.plant_type.astype('category').cat.codes

Посмотрим на данные city.

plants.city.unique()

Преобразуем их аналогично clients.

plants = plants.merge(city_ids, how='left', on='city')\
                           .drop(columns='city')\
                           .rename(columns={'num':'city'})

Проверим в transactions наличие данных по всем магазинам:

missing_plants = transactions[~transactions.plant.isin(plants.plant)].plant.unique()
print(f"Магазины с неизвестными данными:{missing_plants}")
print(f"Количество транзакций в неизвестных магазинах:{transactions[transactions.plant.isin(missing_plants)].plant.count()}")

Результат обработки:

transactions.info()

plants.info()

# 3. Сохранение данных

Сохраним финальные данные в формате pickle для последующего использования. 

path = str(workdir+'/data')
if not os.path.isdir(path):
    os.mkdir(path)
    print('Папка успешно создана!')
else:
    print('Папка уже существует')

transactions.to_pickle(path+'/transactions.pkl')
clients.to_pickle(path+'/clients.pkl')
materials.to_pickle(path+'/materials.pkl')
plants.to_pickle(path+'/plants.pkl')

Сведения о соответствии изначальных идентификаторов категорий и с их целочисленным форматом не требуются. В рамках данной задачи нет необходимости в возврате к первоначальным идентификаторам. Однако, на случай необходимости оставим в данном блокноте таблицу с идентификаторами городов.

city_ids
